feat(passive-voice-sentences-percentage)
* export PassiveVoiceSentencesPercentage to new SentencesAnalyzer class
* change PassiveVoiceSentencesPercentage return type to float
* export extended types strings to constants
* add *n't to be forms to data file
* rename ParagraphAnalyzer to ParagraphsAnalyzer

+++ Crawler/Analyzers/Content/ContentAnalysisResult.cs
@@ -57,7 +57,7 @@ namespace Crawler.Analyzers.Content
         public float NinthDecileBetweenPunctuation { get; set; }
 
         [Result("Passive voice percentage")]
-        public double PassiveVoiceSentencesPercentage { get; set; }
+        public float PassiveVoiceSentencesPercentage { get; set; }
 
         [Normalize]
         [Result("Amount of '?'")]
